pyproject.toml
[tool.poetry]
name = "Curent"
version = "0.1.0"
description = "This repo contains a stream library for integrating with different stream services."
authors = ["Burak Demirtas <burak.demirtas@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^18.2"

[tool.poetry.dev-dependencies]
pytest = "^4.3"
pytest-cover = "^3.0"
pytest-asyncio = "^0.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
